split test_io into io dir
