+++ py-polars/tests/io/.gitignore
@@ -1,0 +1,5 @@
+*.parquet
+*.csv
+*.avro
+*.ipc
+*.json
