+++ py-polars/tests/io/test_avro.py
@@ -1,0 +1,42 @@
+import os
+from typing import List, Literal
+
+import pytest
+
+import polars as pl
+from polars import io
+
+
+@pytest.fixture
+def example_df() -> pl.DataFrame:
+    return pl.DataFrame({"i64": [1, 2], "f64": [0.1, 0.2], "utf8": ["a", "b"]})
+
+
+Compression = Literal["uncompressed", "snappy", "deflate"]
+
+
+@pytest.fixture
+def compressions() -> List[Compression]:
+    return ["uncompressed", "snappy", "deflate"]
+
+
+def test_from_to_buffer(
+    example_df: pl.DataFrame, compressions: List[Compression]
+) -> None:
+    for compression in compressions:
+        buf = io.BytesIO()
+        example_df.to_avro(buf, compression=compression)
+        buf.seek(0)
+        read_df = pl.read_avro(buf)
+        assert example_df.frame_equal(read_df)
+
+
+def test_from_to_file(
+    io_test_dir: str, example_df: pl.DataFrame, compressions: List[Compression]
+) -> None:
+    f = os.path.join(io_test_dir, "small.avro")
+
+    for compression in compressions:
+        example_df.to_avro(f, compression=compression)
+        df_read = pl.read_avro(str(f))
+        assert example_df.frame_equal(df_read)
