+++ py-polars/tests/io/test_pickle.py
@@ -1,0 +1,13 @@
+import pickle
+import polars as pl
+
+
+def test_pickle() -> None:
+    a = pl.Series("a", [1, 2])
+    b = pickle.dumps(a)
+    out = pickle.loads(b)
+    assert a.series_equal(out)
+    df = pl.DataFrame({"a": [1, 2], "b": ["a", None], "c": [True, False]})
+    b = pickle.dumps(df)
+    out = pickle.loads(b)
+    assert df.frame_equal(out, null_equal=True)
