--- py-polars/tests/lazy_io/.gitignore
@@ -1,2 +1,0 @@
-*.parquet
-*.csv
